Conflicts:
	src/gui/kernel/qsimpledrag.cpp
	src/gui/kernel/qwindowsysteminterface.cpp
	src/gui/kernel/qwindowsysteminterface_p.h
	src/plugins/platforms/xcb/qxcbwindow.cpp

+++ src/3rdparty/md4/md4.cpp
@@ -180,7 +180,7 @@ static void md4_update(struct md4_context *ctx, const unsigned char *data, size_
 	saved_lo = ctx->lo;
 	if ((ctx->lo = (saved_lo + size) & 0x1fffffff) < saved_lo)
 		ctx->hi++;
-	ctx->hi += size >> 29;
+	ctx->hi += (quint32)(size >> 29);
 
 	used = saved_lo & 0x3f;
